// src/meta-data/enums/operators/operators.enum.ts
export enum Operators {
    IN = 'in',
    NOT_IN = 'not_in',
    EQUAL = 'equals',
    NOT_EQUALS = 'not_equals',
    LT = 'lt',
    LTE = 'lte',
    GT = 'gt',
    GTE = 'gte',
    CONTAINS = 'contains',
    NOT_CONTAINS = 'not_contains',
    BEFORE = 'before',
    AFTER = 'after',
    BETWEEN = 'between',
    ON = 'on',
    STARTS_WITH = 'starts_with',
    ENDS_WITH = 'ends_with',
    NOT_IN_ANY = 'not_in_any',
    IS_LESS = 'is_less',
    IS_MORE = 'is_more',
    IS_AFTER_NEXT = 'is_after_next',
    IS_BEFORE_LAST = 'is_before_last',
    IS_BLANK = 'is_blank',
    NOT_BLANK = 'not_blank',
}